requirements.txt
flask
flask-smorest
python-dotenv
flask-jwt-extended
flask_cors
mysql-connector-python
twilio